+++ src/Inkluzitron/Extensions/UserExtensions.cs
@@ -1,22 +1,12 @@
 ﻿using Discord;
-using System.Net;
-using System.Threading.Tasks;
 
 namespace Inkluzitron.Extensions
 {
     static public class UserExtensions
     {
-        static public string GetUserAvatar(this IUser user, ImageFormat imageFormat = ImageFormat.Auto, ushort size = 128)
+        static public string GetUserOrDefaultAvatarUrl(this IUser user, ImageFormat imageFormat = ImageFormat.Auto, ushort size = 128)
         {
             return user.GetAvatarUrl(imageFormat, size) ?? user.GetDefaultAvatarUrl();
-        }
-
-        static public async Task<byte[]> DownloadProfilePictureAsync(this IUser user, ImageFormat imageFormat = ImageFormat.Auto, ushort size = 128)
-        {
-            var avatarUrl = user.GetAvatarUrl(imageFormat, size) ?? user.GetDefaultAvatarUrl();
-
-            using var client = new WebClient();
-            return await client.DownloadDataTaskAsync(avatarUrl);
         }
     }
 }
