+++ src/Inkluzitron/appsettings.json
@@ -62,7 +62,5 @@
         "EnableBeersOnTap": true
     },
     "RoleEmoteGuildId": 837628220824223795,
-    "ImageCache": {
-        "DirectoryPath": "./ImageCache"
-    }
+    "CacheDirectoryPath": "./ImageCache"
 }
